Atualizacao da lista 3 e lista de Vetor Extra

+++ Lista 3/Lista3_exe15/main.c
@@ -5,11 +5,11 @@
 int main()
 {
     setlocale(LC_ALL, "Portuguese");
-    int numero, fatorial=1;
+    double numero, fatorial=1;
     printf("Digite um número para ser calculado o valor de seu fatorial: ");
-    scanf("%d", &numero);
-    for (int i=numero; i>=1; i--)
+    scanf("%lf", &numero);
+    for (double i=numero; i>=1; i--)
         fatorial = fatorial *i;
-    printf("O valor do fatorial é: %d ", fatorial);
+    printf("O valor do fatorial é: %lf ", fatorial);
 
 }
